Migrate spacer style from TT1 theme
The following is to replicate the styles found in /twentytwentyone/assets/sass/05-blocks/spacer/_style.scss

Where when in mobile width and the spacer has a style attribute assigned (assumingly for a height value)
	force the height of the spacer to whatever the global spacing unit is set to be.

CSS variables cannot be used for media queries

(an example that does not work...)
@media only screen and (max-width: var(--wp--custom--spacing--mobile-breakpoint)) {

I don't know of a configurable way to manage breakpoints with theme.json

I attempted to construct a possible implementation of those settings in the theme.json
file and couldn't figure out how it might work...

What are we trying to define and allow a user to configure... and what do we just want to be a part of the theme's style?

The fact that the spacer collapses?
At what breakpoint it collapses?
What height it should collapse to?

Or is the below just an aspect of the theme, the already configured spacing unit is enough variance.
And if that is the case how do we normalize the breakpoints?